Visual Studio project file was removed and CMakeLists.txt was added

+++ Headers/DataTypes.h
@@ -5,17 +5,15 @@
 #include <list>
 #include <memory>
 #include <string>
-//#include <vector>
 
 struct TUid
 {
-  uint32_t m_lineIndex;
-  uint32_t m_pos;
+  size_t m_lineIndex;
+  size_t m_pos;
 };
 
 struct TRawData
 {
-  //std::vector<char> m_data;
   std::string m_data;
   TUid m_uid = {0, 0};
   bool m_eol = true;
@@ -26,11 +24,11 @@ struct TResult
 {
   struct TLikeness
   {
-    uint32_t m_pos;
+    size_t m_pos;
     std::string m_subString;
   };
 
-  uint32_t m_line;
+  size_t m_line;
   std::list<TLikeness> m_results;
 };
 using PResult = std::shared_ptr<TResult>;
